KB_Deepspeed/requirement_prv.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_kmp_llvm
abseil-cpp=20211102.0=h27087fc_1
absl-py=1.2.0=pypi_0
accelerate=0.12.0=pyhd8ed1ab_0
aiohttp=3.8.1=py310h5764c6d_1
aiosignal=1.2.0=pyhd8ed1ab_0
arrow-cpp=8.0.0=py310h3098874_0
asttokens=2.0.8=pypi_0
astunparse=1.6.3=pypi_0
async-timeout=4.0.2=pyhd8ed1ab_0
attrs=22.1.0=pyh71513ae_1
aws-c-common=0.4.57=he6710b0_1
aws-c-event-stream=0.1.6=h2531618_5
aws-checksums=0.1.9=he6710b0_0
aws-sdk-cpp=1.8.185=hce553d0_0
bert-score=0.3.11=pypi_0
blas=1.0=mkl
bleurt=0.0.2=pypi_0
boost-cpp=1.78.0=he72f1d9_0
bottleneck=1.3.5=py310ha9d4c09_0
brotli=1.0.9=h166bdaf_7
brotli-bin=1.0.9=h166bdaf_7
brotlipy=0.7.0=py310h7f8727e_1002
bzip2=1.0.8=h7b6447c_0
c-ares=1.18.1=h7f98852_0
ca-certificates=2022.9.24=ha878542_0
cachetools=5.2.0=pypi_0
certifi=2022.9.24=pyhd8ed1ab_0
cffi=1.15.1=py310h74dc2b5_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.1.3=py310hff52083_0
colorama=0.4.5=pyhd8ed1ab_0
cryptography=37.0.1=py310h9ce1e76_0
cudatoolkit=10.2.89=hfd86e86_1
cycler=0.11.0=pypi_0
dataclasses=0.8=pyhc8e2a94_3
datasets=2.4.0=pyhd8ed1ab_0
deepspeed=0.7.6=pypi_0
dill=0.3.5.1=pyhd8ed1ab_0
evaluate=0.2.2=pypi_0
evaluation=0.0.2=pypi_0
executing=0.10.0=pypi_0
ffmpeg=4.3=hf484d3e_0
filelock=3.7.1=pyhd8ed1ab_0
flatbuffers=1.12=pypi_0
fonttools=4.36.0=pypi_0
freetype=2.11.0=h70c0345_0
frozenlist=1.2.0=py310h7f8727e_1
fsspec=2022.7.1=pyhd8ed1ab_0
g2pk=0.9.4=pypi_0
gast=0.4.0=pypi_0
gflags=2.2.2=he1b5a44_1004
giflib=5.2.1=h7b6447c_0
glog=0.3.1=pypi_0
gmp=6.2.1=h295c915_3
gnutls=3.6.15=he1e5248_0
google-auth=2.10.0=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
google-pasta=0.2.0=pypi_0
gputil=1.4.0=pyh9f0ad1d_0
grpc-cpp=1.46.1=h33aed49_0
grpcio=1.47.0=pypi_0
h5py=3.7.0=pypi_0
hjson=3.1.0=pypi_0
huggingface_hub=0.8.1=pyhd8ed1ab_0
icu=70.1=h27087fc_0
idna=3.3=pyhd3eb1b0_0
importlib-metadata=4.11.4=py310hff52083_0
importlib_metadata=4.11.4=hd8ed1ab_0
iniconfig=1.1.1=pypi_0
intel-openmp=2021.4.0=h06a4308_3561
jamo=0.4.1=pypi_0
joblib=1.1.0=pyhd8ed1ab_0
jpeg=9e=h7f8727e_0
jpype1=1.4.0=pypi_0
keras=2.9.0=pypi_0
keras-preprocessing=1.1.2=pypi_0
keyutils=1.6.1=h166bdaf_0
kiwisolver=1.4.4=pypi_0
konlpy=0.6.0=pypi_0
krb5=1.19.3=h3790be6_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
libbrotlicommon=1.0.9=h166bdaf_7
libbrotlidec=1.0.9=h166bdaf_7
libbrotlienc=1.0.9=h166bdaf_7
libclang=14.0.6=pypi_0
libcurl=7.84.0=h91b91d3_0
libedit=3.1.20191231=he28a2e2_2
libev=4.33=h516909a_1
libevent=2.1.10=h9b69904_4
libffi=3.3=he6710b0_2
libgcc-ng=12.2.0=h65d4601_19
libiconv=1.16=h7f8727e_2
libidn2=2.3.2=h7f8727e_0
libnghttp2=1.46.0=hce63b2e_0
libpng=1.6.37=hbc83047_0
libprotobuf=3.20.1=h4ff587b_0
libssh2=1.10.0=ha56f1ee_2
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.16.0=h27cfd23_0
libthrift=0.15.0=he6d91bd_0
libtiff=4.2.0=h2818925_1
libunistring=0.9.10=h27cfd23_0
libuuid=1.0.3=h7f8727e_2
libwebp=1.2.2=h55f646e_0
libwebp-base=1.2.2=h7f8727e_0
littleutils=0.2.2=pypi_0
llvm-openmp=14.0.6=h9e868ea_0
lxml=4.9.1=pypi_0
lz4-c=1.9.3=h295c915_1
markdown=3.4.1=pypi_0
markupsafe=2.1.1=pypi_0
matplotlib=3.5.3=pypi_0
mecab-ko=0.996=h4bd325d_5
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py310h7f8727e_0
mkl_fft=1.3.1=py310hd6ae3a3_0
mkl_random=1.2.2=py310h00e6091_0
mpi=1.0=mpich
mpi4py=3.1.4=py310h37cc914_0
mpich=4.0.3=external_0
multidict=6.0.2=py310h5764c6d_1
multiprocess=0.70.12.2=py310h5764c6d_2
ncurses=6.3=h5eee18b_3
nettle=3.7.3=hbbd107a_1
ninja=1.11.1=pypi_0
nltk=3.7=pyhd3eb1b0_0
numexpr=2.8.3=py310hcea2de6_0
numpy=1.23.1=py310h1794996_0
numpy-base=1.23.1=py310hcba007f_0
nvidia-ml-py3=7.352.0=pypi_0
nvidia-smi=0.1.3=pypi_0
oauthlib=3.2.0=pypi_0
openh264=2.1.1=h4ff587b_0
openssl=1.1.1s=h0b41bf4_1
opt-einsum=3.3.0=pypi_0
orc=1.7.4=h07ed6aa_0
os=0.1.4=0
packaging=21.3=pyhd8ed1ab_0
pandas=1.4.3=py310h6a678d5_0
pillow=9.2.0=py310hace64e9_1
pip=22.3.1=pypi_0
pluggy=1.0.0=pypi_0
protobuf=3.19.4=pypi_0
psutil=5.9.1=py310h5764c6d_0
py=1.11.0=pypi_0
py-cpuinfo=9.0.0=pypi_0
pyarrow=8.0.0=py310h468efa6_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pybind11=2.9.2=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pydantic=1.10.2=pypi_0
pyopenssl=22.0.0=pyhd3eb1b0_0
pyparsing=3.0.9=pyhd8ed1ab_0
pysocks=1.7.1=py310h06a4308_0
pytest=7.1.2=pypi_0
python=3.10.4=h12debd9_0
python-dateutil=2.8.2=pyhd8ed1ab_0
python-gflags=3.1.2=pypi_0
python-mecab-ko=1.0.14=pypi_0
python-xxhash=3.0.0=py310h5764c6d_1
python_abi=3.10=2_cp310
pytorch=1.12.1=py3.10_cuda10.2_cudnn7.6.5_0
pytorch-mutex=1.0=cuda
pytz=2022.1=pyhd8ed1ab_0
pyyaml=6.0=py310h5764c6d_4
qutil=3.2.1=6
re2=2022.04.01=h27087fc_0
readline=8.1.2=h7f8727e_1
regex=2022.7.9=py310h5eee18b_0
requests=2.28.1=py310h06a4308_0
requests-oauthlib=1.3.1=pypi_0
responses=0.18.0=pyhd8ed1ab_0
rsa=4.9=pypi_0
sacremoses=0.0.53=pyhd8ed1ab_0
scipy=1.9.0=pypi_0
sentencepiece=0.1.96=pypi_0
setuptools=61.2.0=py310h06a4308_0
six=1.16.0=pyhd3eb1b0_1
snappy=1.1.9=hbd366e4_1
sorcery=0.2.2=pypi_0
sqlite=3.39.2=h5082296_0
tensorboard=2.9.1=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
tensorflow=2.9.1=pypi_0
tensorflow-estimator=2.9.0=pypi_0
tensorflow-io-gcs-filesystem=0.26.0=pypi_0
termcolor=1.1.0=pypi_0
tf-slim=1.1.0=pypi_0
tk=8.6.12=h1ccaba5_0
tokenizers=0.11.4=py310h3dcd8bd_1
tomli=2.0.1=pypi_0
torchaudio=0.12.1=py310_cu102
torchvision=0.13.1=py310_cu102
tqdm=4.64.0=pyhd8ed1ab_0
transformers=4.21.0=pyhd8ed1ab_0
typing-extensions=4.3.0=py310h06a4308_0
typing_extensions=4.3.0=py310h06a4308_0
tzdata=2022a=hda174b7_0
urllib3=1.26.11=py310h06a4308_0
utf8proc=2.6.1=h27cfd23_0
werkzeug=2.2.2=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
wrapt=1.14.1=pypi_0
xxhash=0.8.0=h7f98852_3
xz=5.2.5=h7f8727e_1
yaml=0.2.5=h7f98852_2
yarl=1.7.2=py310h5764c6d_2
zipp=3.8.1=pyhd8ed1ab_0
zlib=1.2.12=h7f8727e_2
zstd=1.5.2=ha4553b6_0
